requirements.txt
# Python package requirements
discord.py
python-dotenv
aiofiles
asyncio
pycolorise
pydantic
sqlalchemy